int id -> Long id

+++ src/main/java/com/github/plugatarev/computerstore/models/DesktopComputerDTO.java
@@ -11,7 +11,7 @@ public class DesktopComputerDTO extends ProductDTO {
     @NotEmpty
     private final DesktopForm formFactor;
 
-    public DesktopComputerDTO(int id, @NotEmpty String serialNumber, @NotEmpty Double price, @NotEmpty String vendor,
+    public DesktopComputerDTO(Long id, @NotEmpty String serialNumber, @NotEmpty Double price, @NotEmpty String vendor,
                               @NotEmpty Integer productAvailable, @NotEmpty ProductType type, @NotEmpty DesktopForm formFactor) {
         super(id, serialNumber, price, vendor, productAvailable, type);
         this.formFactor = formFactor;
