+++ src/main/java/com/github/plugatarev/computerstore/models/MonitorDTO.java
@@ -10,7 +10,7 @@ public class MonitorDTO extends ProductDTO {
     @NotEmpty
     private final Double diagonal;
 
-    public MonitorDTO(Integer id, @NotEmpty String serialNumber, @NotEmpty Double price, @NotEmpty String vendor,
+    public MonitorDTO(Long id, @NotEmpty String serialNumber, @NotEmpty Double price, @NotEmpty String vendor,
                       Integer productAvailable, @NotEmpty ProductType type, @NotEmpty Double diagonal) {
         super(id, serialNumber, price, vendor, productAvailable, type);
         this.diagonal = diagonal;
